CC/app/requirements.txt
Flask==3.0.0
keras==2.15.0
numpy==1.26.2
Pillow==10.1.0
Pillow==10.1.0
waitress==2.1.2
